- refactored camera handling and frame settings into the CameraSettings scene extension
