+++ packages/studio-base/src/panels/ThreeDeeRender/Renderer.ts
@@ -6,7 +6,6 @@ import EventEmitter from "eventemitter3";
 import i18next from "i18next";
 import { Immutable, produce } from "immer";
 import * as THREE from "three";
-import { OrbitControls } from "three/examples/jsm/controls/OrbitControls";
 import { DeepPartial } from "ts-essentials";
 import { v4 as uuidv4 } from "uuid";
 
@@ -50,7 +49,6 @@ import {
 import { CameraStateSettings } from "./renderables/CameraStateSettings";
 import { Cameras } from "./renderables/Cameras";
 import { FrameAxes, LayerSettingsTransform } from "./renderables/FrameAxes";
-import { FrameSettings } from "./renderables/FrameSettings";
 import { Grids } from "./renderables/Grids";
 import { ImageMode } from "./renderables/ImageMode";
 import { Images } from "./renderables/Images";
@@ -80,14 +78,7 @@ import {
   Vector3,
 } from "./ros";
 import { BaseSettings, CustomLayerSettings, SelectEntry } from "./settings";
-import {
-  AddTransformResult,
-  CoordinateFrame,
-  makePose,
-  Pose,
-  Transform,
-  TransformTree,
-} from "./transforms";
+import { AddTransformResult, CoordinateFrame, Transform, TransformTree } from "./transforms";
 import { InterfaceMode } from "./types";
 
 const log = Logger.getLogger(__filename);
@@ -249,10 +240,6 @@ const DARK_BACKDROP = new THREE.Color(dark.background?.default);
 const LAYER_DEFAULT = 0;
 const LAYER_SELECTED = 1;
 
-const UNIT_X = new THREE.Vector3(1, 0, 0);
-const UNIT_Z = new THREE.Vector3(0, 0, 1);
-const PI_2 = Math.PI / 2;
-
 // Coordinate frames named in [REP-105](https://www.ros.org/reps/rep-0105.html)
 const DEFAULT_FRAME_IDS = ["base_link", "odom", "map", "earth"];
 
@@ -267,13 +254,7 @@ const CYCLE_DETECTED = "CYCLE_DETECTED";
 const RENDERER_ID = "foxglove.Renderer";
 
 const tempColor = new THREE.Color();
-const tempVec3 = new THREE.Vector3();
 const tempVec2 = new THREE.Vector2();
-const tempSpherical = new THREE.Spherical();
-const tempEuler = new THREE.Euler();
-
-// used for holding unfollowPoseSnapshot in render frame every new frame
-const snapshotInRenderFrame = makePose();
 
 // We use a patched version of THREE.js where the internal WebGLShaderCache class has been
 // modified to allow caching based on `vertexShaderKey` and/or `fragmentShaderKey` instead of
@@ -340,18 +321,6 @@ export class Renderer extends EventEmitter<RendererEvents> {
   public measurementTool: MeasurementTool;
   public publishClickTool: PublishClickTool;
 
-  private perspectiveCamera: THREE.PerspectiveCamera;
-  private orthographicCamera: THREE.OrthographicCamera;
-  // This group is used to transform the cameras based on the Frame follow mode
-  // quaternion is affected in stationary and position-only follow modes
-  // both position and quaternion of the group are affected in stationary mode
-  private cameraGroup: THREE.Group;
-  private aspect: number;
-  private controls: OrbitControls;
-  public followMode: FollowMode;
-  // The pose of the render frame in the fixed frame when following was disabled
-  public unfollowPoseSnapshot: Pose | undefined;
-
   // Are we connected to a ROS data source? Normalize coordinate frames if so by
   // stripping any leading "/" prefix. See `normalizeFrameId()` for details.
   public ros = false;
@@ -374,7 +343,6 @@ export class Renderer extends EventEmitter<RendererEvents> {
 
   private _prevResolution = new THREE.Vector2();
   private _pickingEnabled = false;
-  private _isUpdatingCameraState = false;
   private _animationFrame?: number;
   private _cameraSyncError: undefined | string;
   private _devicePixelRatioMediaQuery?: MediaQueryList;
@@ -450,31 +418,6 @@ export class Renderer extends EventEmitter<RendererEvents> {
     this.scene.add(this.dirLight);
     this.scene.add(this.hemiLight);
 
-    this.perspectiveCamera = new THREE.PerspectiveCamera();
-    this.orthographicCamera = new THREE.OrthographicCamera();
-    this.cameraGroup = new THREE.Group();
-
-    this.cameraGroup.add(this.perspectiveCamera);
-    this.cameraGroup.add(this.orthographicCamera);
-    this.scene.add(this.cameraGroup);
-
-    this.controls = new OrbitControls(this.perspectiveCamera, this.canvas);
-    this.controls.screenSpacePanning = false; // only allow panning in the XY plane
-    this.controls.mouseButtons.LEFT = THREE.MOUSE.PAN;
-    this.controls.mouseButtons.RIGHT = THREE.MOUSE.ROTATE;
-    this.controls.touches.ONE = THREE.TOUCH.PAN;
-    this.controls.touches.TWO = THREE.TOUCH.DOLLY_ROTATE;
-    this.controls.addEventListener("change", () => {
-      if (!this._isUpdatingCameraState) {
-        this.emit("cameraMove", this);
-      }
-    });
-
-    // Make the canvas able to receive keyboard events and setup WASD controls
-    canvas.tabIndex = 1000;
-    this.controls.keys = { LEFT: "KeyA", RIGHT: "KeyD", UP: "KeyW", BOTTOM: "KeyS" };
-    this.controls.listenToKeyEvents(canvas);
-
     this.input = new Input(canvas, () => this.activeCamera());
     this.input.on("resize", (size) => this.resizeHandler(size));
     this.input.on("click", (cursorCoords) => this.clickHandler(cursorCoords));
@@ -486,16 +429,14 @@ export class Renderer extends EventEmitter<RendererEvents> {
     this.scene.add(this.selectionBackdrop);
 
     this.followFrameId = config.followTf;
-    this.followMode = config.followMode;
 
     const samples = msaaSamples(this.gl.capabilities);
     const renderSize = this.gl.getDrawingBufferSize(tempVec2);
-    this.aspect = renderSize.width / renderSize.height;
     log.debug(`Initialized ${renderSize.width}x${renderSize.height} renderer (${samples}x MSAA)`);
 
     this.measurementTool = new MeasurementTool(this);
     this.publishClickTool = new PublishClickTool(this);
-    this.cameraStateSettings = new CameraStateSettings(this);
+    this.cameraStateSettings = new CameraStateSettings(this, this.canvas, renderSize);
 
     // Internal handlers for TF messages to update the transform tree
     this.addSchemaSubscriptions(FRAME_TRANSFORM_DATATYPES, {
@@ -526,7 +467,6 @@ export class Renderer extends EventEmitter<RendererEvents> {
       case "3d":
         this.addSceneExtension(this.cameraStateSettings);
         this.addSceneExtension(new PublishSettings(this));
-        this.addSceneExtension(new FrameSettings(this));
         break;
     }
 
@@ -551,7 +491,7 @@ export class Renderer extends EventEmitter<RendererEvents> {
 
     this._watchDevicePixelRatio();
 
-    this._updateCameras(config.cameraState);
+    this.setCameraState(config.cameraState);
     this.animationFrame();
   }
 
@@ -577,8 +517,6 @@ export class Renderer extends EventEmitter<RendererEvents> {
 
     this.settings.removeAllListeners();
     this.input.removeAllListeners();
-
-    this.controls.dispose();
 
     for (const extension of this.sceneExtensions.values()) {
       extension.dispose();
@@ -946,78 +884,12 @@ export class Renderer extends EventEmitter<RendererEvents> {
     this.settings.setLabel(["layers"], label);
   }
 
-  /** Translate a CameraState to the three.js coordinate system */
-  private _updateCameras(cameraState: CameraState): void {
-    const targetOffset = tempVec3;
-    targetOffset.fromArray(cameraState.targetOffset);
-
-    const phi = THREE.MathUtils.degToRad(cameraState.phi);
-    const theta = -THREE.MathUtils.degToRad(cameraState.thetaOffset);
-
-    // Always update the perspective camera even if the current mode is orthographic. This is needed
-    // to make the OrbitControls work properly since they track the perspective camera.
-    // https://github.com/foxglove/studio/issues/4138
-
-    // Convert the camera spherical coordinates (radius, phi, theta) to Cartesian (X, Y, Z)
-    tempSpherical.set(cameraState.distance, phi, theta);
-    this.perspectiveCamera.position.setFromSpherical(tempSpherical).applyAxisAngle(UNIT_X, PI_2);
-    this.perspectiveCamera.position.add(targetOffset);
-
-    // Convert the camera spherical coordinates (phi, theta) to a quaternion rotation
-    this.perspectiveCamera.quaternion.setFromEuler(tempEuler.set(phi, 0, theta, "ZYX"));
-    this.perspectiveCamera.fov = cameraState.fovy;
-    this.perspectiveCamera.near = cameraState.near;
-    this.perspectiveCamera.far = cameraState.far;
-    this.perspectiveCamera.aspect = this.aspect;
-    this.perspectiveCamera.updateProjectionMatrix();
-
-    this.controls.target.copy(targetOffset);
-
-    if (cameraState.perspective) {
-      // Unlock the polar angle (pitch axis)
-      this.controls.minPolarAngle = 0;
-      this.controls.maxPolarAngle = Math.PI;
-    } else {
-      // Lock the polar angle during 2D mode
-      const curPolarAngle = THREE.MathUtils.degToRad(this.config.cameraState.phi);
-      this.controls.minPolarAngle = this.controls.maxPolarAngle = curPolarAngle;
-
-      this.orthographicCamera.position.set(targetOffset.x, targetOffset.y, cameraState.far / 2);
-      this.orthographicCamera.quaternion.setFromAxisAngle(UNIT_Z, theta);
-      this.orthographicCamera.left = (-cameraState.distance / 2) * this.aspect;
-      this.orthographicCamera.right = (cameraState.distance / 2) * this.aspect;
-      this.orthographicCamera.top = cameraState.distance / 2;
-      this.orthographicCamera.bottom = -cameraState.distance / 2;
-      this.orthographicCamera.near = cameraState.near;
-      this.orthographicCamera.far = cameraState.far;
-      this.orthographicCamera.updateProjectionMatrix();
-    }
-  }
-
   public setCameraState(cameraState: CameraState): void {
-    this._isUpdatingCameraState = true;
-    this._updateCameras(cameraState);
-    // only active for follow pose mode because it introduces strange behavior into the other modes
-    // due to the fact that they are manipulating the camera after update with the `cameraGroup`
-    if (this.followMode === "follow-pose") {
-      this.controls.update();
-    }
-    this._isUpdatingCameraState = false;
+    this.cameraStateSettings.setCameraState(cameraState);
   }
 
   public getCameraState(): CameraState {
-    return {
-      perspective: this.config.cameraState.perspective,
-      distance: this.controls.getDistance(),
-      phi: THREE.MathUtils.radToDeg(this.controls.getPolarAngle()),
-      thetaOffset: THREE.MathUtils.radToDeg(-this.controls.getAzimuthalAngle()),
-      targetOffset: [this.controls.target.x, this.controls.target.y, this.controls.target.z],
-      target: this.config.cameraState.target,
-      targetOrientation: this.config.cameraState.targetOrientation,
-      fovy: this.config.cameraState.fovy,
-      near: this.config.cameraState.near,
-      far: this.config.cameraState.far,
-    };
+    return this.cameraStateSettings.getCameraState();
   }
 
   public setSelectedRenderable(selection: PickedRenderable | undefined): void {
@@ -1050,7 +922,7 @@ export class Renderer extends EventEmitter<RendererEvents> {
   }
 
   private activeCamera(): THREE.PerspectiveCamera | THREE.OrthographicCamera {
-    return this.config.cameraState.perspective ? this.perspectiveCamera : this.orthographicCamera;
+    return this.cameraStateSettings.getActiveCamera();
   }
 
   public addMessageEvent(messageEvent: Readonly<MessageEvent<unknown>>): void {
@@ -1202,7 +1074,7 @@ export class Renderer extends EventEmitter<RendererEvents> {
 
   private frameHandler = (currentTime: bigint): void => {
     this.currentTime = currentTime;
-    this._updateFrames(currentTime);
+    this._updateFrames();
     this._updateResolution();
 
     this.gl.clear();
@@ -1215,48 +1087,6 @@ export class Renderer extends EventEmitter<RendererEvents> {
     // use the FALLBACK_FRAME_ID if renderFrame is undefined and there are no options for transforms
     const renderFrameId = this.renderFrameId ?? CoordinateFrame.FALLBACK_FRAME_ID;
     const fixedFrameId = this.fixedFrameId ?? CoordinateFrame.FALLBACK_FRAME_ID;
-
-    const renderFrame = this.transformTree.frame(renderFrameId);
-    const fixedFrame = this.transformTree.frame(fixedFrameId);
-
-    // If in stationary or follow-position modes
-    if (
-      this.followMode !== "follow-pose" &&
-      this.unfollowPoseSnapshot &&
-      renderFrame &&
-      fixedFrame
-    ) {
-      renderFrame.applyLocal(
-        snapshotInRenderFrame,
-        this.unfollowPoseSnapshot,
-        fixedFrame,
-        currentTime,
-      );
-      /**
-       * the application of the unfollowPoseSnapshot position and orientation
-       * components makes the camera position and rotation static relative to the fixed frame.
-       * So when the display frame changes the angle of the camera relative
-       * to the scene will not change because only the snapshotPose orientation is applied
-       */
-      if (this.followMode === "follow-position") {
-        // only make orientation static/stationary in this mode
-        // the position still follows the frame
-        this.cameraGroup.position.set(0, 0, 0);
-      } else {
-        this.cameraGroup.position.set(
-          snapshotInRenderFrame.position.x,
-          snapshotInRenderFrame.position.y,
-          snapshotInRenderFrame.position.z,
-        );
-      }
-      // this negates the rotation of the changes in renderFrame
-      this.cameraGroup.quaternion.set(
-        snapshotInRenderFrame.orientation.x,
-        snapshotInRenderFrame.orientation.y,
-        snapshotInRenderFrame.orientation.z,
-        snapshotInRenderFrame.orientation.w,
-      );
-    }
 
     for (const sceneExtension of this.sceneExtensions.values()) {
       sceneExtension.startFrame(currentTime, renderFrameId, fixedFrameId);
@@ -1281,8 +1111,7 @@ export class Renderer extends EventEmitter<RendererEvents> {
     this.gl.setSize(size.width, size.height);
 
     const renderSize = this.gl.getDrawingBufferSize(tempVec2);
-    this.aspect = renderSize.width / renderSize.height;
-    this._updateCameras(this.config.cameraState);
+    this.cameraStateSettings.handleResize(renderSize);
 
     log.debug(`Resized renderer to ${renderSize.width}x${renderSize.height}`);
     this.animationFrame();
@@ -1465,7 +1294,7 @@ export class Renderer extends EventEmitter<RendererEvents> {
   /** Tracks the number of frames so we can recompute the defaultFrameId when frames are added. */
   private _lastTransformFrameCount = 0;
 
-  private _updateFrames(currentTime: bigint): void {
+  private _updateFrames(): void {
     if (
       this.followFrameId != undefined &&
       this.renderFrameId !== this.followFrameId &&
@@ -1531,59 +1360,10 @@ export class Renderer extends EventEmitter<RendererEvents> {
       } else {
         log.debug(`Changing fixed frame from "${this.fixedFrameId}" to "${fixedFrameId}"`);
       }
-      // Set the unfollowPoseSnapshot to undefined because there is a new fixed frame for the snapshot
-      // This keeps the camera settings offsets based off of the display frame rather than old fixed frame.
-      this.unfollowPoseSnapshot = undefined;
       this.fixedFrameId = fixedFrameId;
     }
 
-    // Should only occur on reload when the saved followMode is not follow
-    if (this.followMode !== "follow-pose" && !this.unfollowPoseSnapshot) {
-      // Snapshot the current pose of the render frame in the fixed frame
-      this.unfollowPoseSnapshot = makePose();
-      fixedFrame.applyLocal(
-        this.unfollowPoseSnapshot,
-        this.unfollowPoseSnapshot,
-        frame,
-        currentTime,
-      );
-    }
     this.settings.errors.clearPath(FOLLOW_TF_PATH);
-  }
-
-  // This should not be called on initialization only on settings changes
-  public updateFollowMode(newFollowMode: FollowMode): void {
-    if (this.followMode === newFollowMode) {
-      return;
-    }
-
-    if (!this.renderFrameId || !this.fixedFrameId) {
-      this.followMode = newFollowMode;
-      return;
-    }
-
-    const renderFrame = this.transformTree.frame(this.renderFrameId);
-    const fixedFrame = this.transformTree.frame(this.fixedFrameId);
-
-    if (!renderFrame || !fixedFrame) {
-      // if this happens it will be set on initialization in _updateFrames
-      this.followMode = newFollowMode;
-      return;
-    }
-
-    // always create a new snapshot when changing frames to minimize old snapshots causing camera jumps
-    this.unfollowPoseSnapshot = makePose();
-    fixedFrame.applyLocal(
-      this.unfollowPoseSnapshot,
-      this.unfollowPoseSnapshot,
-      renderFrame,
-      this.currentTime,
-    );
-
-    // reset any applied cameraGroup settings so that they aren't applied in follow mode
-    this.cameraGroup.position.set(0, 0, 0);
-    this.cameraGroup.quaternion.set(0, 0, 0, 1);
-    this.followMode = newFollowMode;
   }
 
   private _updateResolution(): void {
